feat: Auto-anchor support for strat detail page
