+++ src/components/StratAside.vue
@@ -1,10 +1,11 @@
 <script setup lang="ts">
 import { motion } from 'motion-v'
-import { computed } from 'vue'
+import { computed, onMounted, onUnmounted } from 'vue'
 
 import type { DutyMechanic, DutyPhase } from '@/data/duty'
 
 import { Separator } from '@/components/shadcn-vue/separator'
+import { useAnchorObserver } from '@/lib/anchor'
 
 const props = defineProps<{
   currentUrl: string
@@ -25,6 +26,21 @@ const mechanics = computed(() => {
 const baseNavItemStyle = 'relative py-1.5 px-2 rounded-md transition-colors duration-300 text-sm flex items-center'
 const activeNavItemStyle = 'bg-primary/10 text-primary font-medium'
 const hoverNavItemStyle = 'hover:bg-muted/50'
+
+const {
+  click: handleNavClick,
+  activate: handleActivateAnchorObserver,
+  deactivate: handleDeactivateAnchorObserver,
+  activeId,
+} = useAnchorObserver(mechanics.value.map(item => item.href?.slice(1)))
+
+onMounted(async () => {
+  await handleActivateAnchorObserver()
+})
+
+onUnmounted(() => {
+  handleDeactivateAnchorObserver()
+})
 </script>
 
 <template>
@@ -69,8 +85,13 @@ const hoverNavItemStyle = 'hover:bg-muted/50'
             :key="item.href"
             :while-hover="{ x: 2 }"
             :transition="{ type: 'tween', duration: 0.2 }"
+            @click.prevent="handleNavClick(item.href.slice(1))"
           >
-            <a :href="item.href" :class="`${baseNavItemStyle} ${hoverNavItemStyle}`">
+            <a
+              :href="item.href"
+              :class="`${baseNavItemStyle} ${hoverNavItemStyle} ${activeId === item.href?.slice(1) ? activeNavItemStyle : ''}`"
+            >
+              <span v-if="activeId === item.href?.slice(1)" class="bg-primary absolute left-0 h-4 w-1 rounded-full" />
               <span class="ml-1">{{ item.name }}</span>
             </a>
           </motion.li>
